Rebuilt the app, implementing two of the three core functions.
The architecture of the app has been completely redesigned to present a more logical flow of navigation through the app. The ability to register a new account, login to an account, or login later (creating an anonymous user), as well as loging out of the account, has all been implemented. Finally, a list of all trips currently running abnormally is displayed, with a selection displaying more detailed information about that particular trip.